(PC-29420) feat(ui): subtitles in childen for HorizontalTile component
